requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.267
openai==0.28.0
chromadb==0.4.6
jinja2==3.1.2
pydantic==2.3.0
gitpython==3.1.32
